Building in Ubuntu

+++ clibs/Compression/LZ4/lz4.c
@@ -746,8 +746,8 @@ int LZ4_uncompress_unknownOutputSize(
         if unlikely(op-ref<STEPSIZE)
         {
 #if LZ4_ARCH64
-            size_t dec2table[]={0, 0, 0, -1, 0, 1, 2, 3};
-            size_t dec2 = dec2table[op-ref];
+            int dec2table[]={0, 0, 0, -1, 0, 1, 2, 3};
+            int dec2 = dec2table[op-ref];
 #else
             const int dec2 = 0;
 #endif
